Polled exchange added to SPIv3.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@11244 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/ports/STM32/LLD/SPIv3/hal_spi_lld.h
@@ -694,7 +694,7 @@ extern "C" {
   void spi_lld_send(SPIDriver *spip, size_t n, const void *txbuf);
   void spi_lld_receive(SPIDriver *spip, size_t n, void *rxbuf);
   void spi_lld_abort(SPIDriver *spip);
-  uint16_t spi_lld_polled_exchange(SPIDriver *spip, uint16_t frame);
+  uint32_t spi_lld_polled_exchange(SPIDriver *spip, uint32_t frame);
 #ifdef __cplusplus
 }
 #endif
